+++ proto-lens-optparse/Changelog.md
@@ -1,8 +1,9 @@
 # Changelog for `proto-lens-optparse`
 
-## Pending
+## v0.1.1.6
 - Bump lower bounds to base-4.10 (ghc-8.2).
 - Support dependencies on base-4.13 (ghc-8.8) and optparse-applicative-0.15.
+- Bump upper bounds for proto-lens-0.6.
 
 
 ## v0.1.1.5